requirements.txt
streamlit==1.28.0
ultralytics==8.0.200
numpy
pillow